Modernize TGeoMgrFromDdd
- consolidated into 1 file in plugins directory
- added consumes use ESGetToken
- added fillDescriptions
