--- Fireworks/Geometry/plugins/TGeoMgrFromDddPlugin.cc
@@ -1,4 +1,0 @@
-#include "Fireworks/Geometry/interface/TGeoMgrFromDdd.h"
-#include "FWCore/Framework/interface/ModuleFactory.h"
-
-DEFINE_FWK_EVENTSETUP_MODULE(TGeoMgrFromDdd);
